Turned TK_DEBUG off.
Fixed (unsigned long) and (long) castings to avoid compiler warnings.

git-svn-id: file:///Users/mdouma46/Developer/sourceFinagler/branches/2.0.3@117 c3d3028b-b5cd-4c35-8691-6fb8b2347b27

+++ TextureKit/TKImage.m
@@ -22,7 +22,7 @@
 #define TK_DEBUG 0
 
 TEXTUREKIT_STATIC_INLINE NSString *TKImageKey(NSUInteger anUInteger) {
-	return [NSString stringWithFormat:@"%lu", anUInteger];
+	return [NSString stringWithFormat:@"%lu", (unsigned long)anUInteger];
 }
 
 
@@ -1073,10 +1073,10 @@ static inline NSString *NSStringFromImageType(TKImageType aType) {
 	NSMutableString *description = [NSMutableString stringWithString:[super description]];
 	[description appendFormat:@"\n"];
 	[description appendFormat:@"imageType == %@\n", NSStringFromImageType(type)];
-	[description appendFormat:@"sliceCount == %lu\n", sliceCount];
-	[description appendFormat:@"faceCount == %lu\n", faceCount];
-	[description appendFormat:@"frameCount == %lu\n", frameCount];
-	[description appendFormat:@"mipmapCount == %lu\n", mipmapCount];
+	[description appendFormat:@"sliceCount == %lu\n", (unsigned long)sliceCount];
+	[description appendFormat:@"faceCount == %lu\n", (unsigned long)faceCount];
+	[description appendFormat:@"frameCount == %lu\n", (unsigned long)frameCount];
+	[description appendFormat:@"mipmapCount == %lu\n", (unsigned long)mipmapCount];
 	[description appendFormat:@"reps == %@", reps];
 	return description;
 }
